Fix some bugs in hr ( 2 )

+++ modules/hr/models/Designation_model.php
@@ -115,6 +115,11 @@ class Designation_model extends App_Model{
     }
 
     public function delete($id, $simpleDelete = false){
+        if (is_reference_in_table('designation', db_prefix() . 'hr_extra_info', $id)) {
+            return [
+                'referenced' => true,
+            ];
+        }
         $this->db->where('id', $id);
         $this->db->delete($this->table_name);
         if ($this->db->affected_rows() > 0) {
