+++ modules/hr/models/Sub_department_model.php
@@ -75,6 +75,11 @@ class Sub_department_model extends App_Model{
     }
 
     public function delete($id, $simpleDelete = false){
+        if (is_reference_in_table('sub_department', db_prefix() . 'hr_extra_info', $id)) {
+            return [
+                'referenced' => true,
+            ];
+        }
         $this->db->where('id', $id);
         $this->db->delete($this->table_name);
         if ($this->db->affected_rows() > 0) {
